+++ tests/acceptance/map-query-params-test.ts
@@ -124,7 +124,11 @@ module('Acceptance | map query params', function (hooks) {
     assert.dom('[data-test-navbar-refresh]').exists();
     assert
       .dom('[data-test-navbar-refresh]')
-      .hasAttribute('title', 'Refresh map and station data (in 2 minutes)');
+      .hasAttribute(
+        'title',
+        'Refresh map and station data (00:00 since last refresh)'
+      );
+    assert.dom('[data-test-navbar-refresh]').hasText('00:00');
 
     await click('[data-test-navbar-refresh]');
 
@@ -134,7 +138,11 @@ module('Acceptance | map query params', function (hooks) {
     );
     assert
       .dom('[data-test-navbar-refresh]')
-      .hasAttribute('title', 'Refresh map and station data (in 2 minutes)');
+      .hasAttribute(
+        'title',
+        'Refresh map and station data (00:00 since last refresh)'
+      );
+    assert.dom('[data-test-navbar-refresh]').hasText('00:00');
   });
 
   test('it auto refreshes stations after the refresh interval', async function (assert) {